Reimplement Aereck
Move Aereck and implement plugin system. Fix conversation Head.
